+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.144] - 2026-01-16
+### Added
+* `BitmaskUtils.Contains` to check if a bit exists
+### Changed
+* `LayerMaskExtensions` to use `BitmaskUtils.Contains` when calculating `ContainsLayer`
+
 ## [0.0.143] - 2026-01-08
 ### Changed
 * `InputManager` to be just `InputManager` and now require extending classes to override the `Instance` property